Made some changes to CPAxis and CPAxisSet. Mostly clean ups of last commit (dirkjan).
One major change was to remove the axis code from CPPlotSpace. I would rather not have a plot space having any knowledge of an axes set. It should be the other way around: an axis holds a pointer to a plot space. We need a cleaner solution for updating the bounds of the axis set, probably at a higher level (eg CPGraph).