Got axis constraints to work, and added an example in CPTestApp.
These constraints implement a springs and struts model to constrain an axis in the plot area. You can have it at a fixed distance from any side, or floating at a fixed ratio between edges (eg centered). You can also pin it to any side of the plot area.

Also made some minor changes, the most significant of which was changing the property constantCoordinateValue to the more descriptive orthogonalCoordinateDecimal.